♻️ validation 파일에서 format함수들 format.js로 분리 #20
